Crate polkadot_cli

Polkadot CLI library.

Re-exports

pub use service;

Structs

Cli

RunCmd

Enums

Error

Error type for the CLI.

Subcommand

Constants

NODE_VERSION

The current node version, which takes the basic SemVer form <major>.<minor>.<patch>. In general, minor should be bumped on every release while major or patch releases are relatively rare.

Traits

CoreApi

The Core runtime api that every Substrate runtime needs to implement.

IdentifyVariant

Can be called for a Configuration to identify which network the configuration targets.

ProvideRuntimeApi

Something that provides a runtime api.

Functions

prepared_overseer_builder

Obtain a prepared OverseerBuilder, that is initialized with all default values.

run

Parses polkadot specific CLI arguments and run the service.

run_node

Launch a node, accepting arguments just like a regular node, accepts an alternative overseer generator, to adjust behavior for integration tests as needed. malus_finality_delay restrict finality votes of this node to be at most best_block - malus_finality_delay height.

Type Definitions

Block

Block type.

BlockId

Block ID.

Result

Result type alias for the CLI.

TFullClient

Full client type.