Crate polkadot_cli

Polkadot CLI library.

Re-exports

pub use service;

Structs

Cli

RunCmd

ValidationWorkerCommand

Enums

Error
Error type for the CLI.

Subcommand

Traits

CoreApi
The Core runtime api that every Substrate runtime needs to implement.

IdentifyVariant
Can be called for a Configuration to identify which network the configuration targets.

ProvideRuntimeApi
Something that provides a runtime api.

RuntimeApiCollection
A set of APIs that polkadot-like runtimes must implement.

Functions

create_default_subsystems
Create a default, unaltered set of subsystems.

run
Parses polkadot specific CLI arguments and run the service.

run_node
Launch a node, accepting arguments just like a regular node, accepts an alternative overseer generator, to adjust behavior for integration tests as needed.

Type Definitions

Block
Block type.

Result
Result type alias for the CLI.

TFullClient
Full client type.